CARDGAME
There is a deck of cards that is shuffled and split equally between 2 players. Each Player gets a deck and then starts the game. To play each player shows the topmost card of their deck. The Player with the highest card wins the round. Ace is the highest card. The winner gets twice his/her bet. A tie implies both Players lose their bet to the House. When a Player is out of cards or the Player out of cash - the game ends. The Player with the most cash is declared winner.